api/pyproject.toml
[project]
name = "api"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.5",
    "numpy>=2.1.3",
    "onnxruntime>=1.20.1",
    "opencv-python>=4.10.0.84",
    "pyserial>=3.5",
    "python-multipart>=0.0.18",
    "uvicorn>=0.32.1",
]
